@@ -1,5 +1,6 @@
 [metadata]
 name = pe
+version = attr: pe.__version__
 description = Library for Parsing Expression Grammars (PEG)
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -10,8 +11,6 @@ author = Michael Wayne Goodman
 author_email = goodman.m.w@gmail.com
 url = https://github.com/goodmami/pe/
 
-requires_python = ">=3.6"
-
 classifiers =
   Development Status :: 4 - Beta
   Environment :: Console
@@ -20,10 +19,10 @@ classifiers =
   Programming Language :: Cython
   Programming Language :: Python
   Programming Language :: Python :: 3
-  Programming Language :: Python :: 3.6
-  Programming Language :: Python :: 3.7
   Programming Language :: Python :: 3.8
   Programming Language :: Python :: 3.9
+  Programming Language :: Python :: 3.10
+  Programming Language :: Python :: 3.11
   Topic :: Software Development :: Libraries :: Python Modules
   Topic :: Text Processing
   Topic :: Utilities
@@ -34,12 +33,12 @@ project_urls =
 
 
 [options]
-python_requires = >= 3.6
+python_requires = >= 3.8
 zip_safe = False
 packages = pe
 test_suite = test
 setup_requires =
-    setuptools >=46.4.0
+    setuptools >= 67.8
     Cython ~= 0.29
 tests_require =
     pytest
